created an interface for the haploytpe caller integraiton tests between spark and non-spark versions
